+++ src/main/java/com/dmu/stock/entity/UserStock.java
@@ -7,6 +7,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -22,9 +23,9 @@ public class UserStock {
 
     private String stockCode; // 종목코드
 
-    private double avgPrice; // 평단가
+    private BigDecimal avgPrice; // 평단가
 
-    private double quantity; //수량
+    private BigDecimal quantity; //수량
 
     @Enumerated(EnumType.STRING)
     private StockType type; //국내, 미국주식 구분
